pyproject.toml
[project]
name = "hack-europe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "anthropic>=0.83.0",
    "fastapi>=0.129.1",
    "greenlet>=3.3.2",
    "jinja2>=3.1.6",
    "linkup-sdk>=0.12.0",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "python-pptx>=1.0.2",
    "sqlmodel>=0.0.35",
    "uvicorn[standard]>=0.41.0",
    "weasyprint>=68.1",
    "websockets>=16.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.2",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
check_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]
